fantasy-football-api/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
boto3==1.34.0